requirements.txt
open-interpreter
playwright
pillow
